+++ lib/canopus/language/symbol.rb
@@ -3,5 +3,6 @@
 module Canopus
   module Language
     Symbol = Data.define(:name, :kind, :range, :selection, :depth)
+    DocumentSymbol = Data.define(:id, :name, :kind, :range, :selection, :depth, :parent_id)
   end
 end