+++ lib/canopus/workspace/language_server_configurable.rb
@@ -96,6 +96,7 @@ module Canopus
         invalidate_diagnostics
         invalidate_inlay_hints(client: client)
         invalidate_code_lenses(client: client)
+        invalidate_sticky_symbols(client: client)
         (@retired_language_clients ||= ObjectSpace::WeakMap.new)[client] = true
         @opened_lsp_documents&.keys&.each do |key|
           next unless key.first.equal?(client)